refactor(layout): rename section directory to page-sections
